Migrated 2 views into 1
