+++ app/src/main/AndroidManifest.xml
@@ -10,11 +10,11 @@
         android:roundIcon="@drawable/ic_app_icon"
         android:supportsRtl="true"
         android:theme="@style/Theme.RepairTool">
-        <activity
-            android:name=".volunteer.VolunteerActivity"
-            android:exported="true"
-            android:label="@string/title_activity_repair"
-            android:theme="@style/Theme.RepairTool.NoActionBar" />
+<!--        <activity-->
+<!--            android:name=".volunteer.VolunteerActivity"-->
+<!--            android:exported="true"-->
+<!--            android:label="@string/title_activity_repair"-->
+<!--            android:theme="@style/Theme.RepairTool.NoActionBar" />-->
         <activity
             android:name=".repairman.RepairmanActivity"
             android:exported="true"
