+++ app/src/main/java/com/example/repairtool/login/LoginView.kt
@@ -26,7 +26,6 @@ import com.example.repairtool.repairman.RepairmanActivity
 import com.example.repairtool.ui.theme.RepairToolTheme
 import com.example.repairtool.utilities.textfield.singelLinePasswordLabel
 import com.example.repairtool.utilities.textfield.singleLineInputLabel
-import com.example.repairtool.volunteer.VolunteerActivity
 
 //Main function to show loginpage
 @Composable
@@ -70,6 +69,14 @@ fun LoginView() {
         Spacer(modifier = Modifier.padding(top = 30.dp))
 
         //Login button AND forgot password button
+        /*
+        Op dit moment moet de navigatie en code die hier voor gebruikt wordt nog gemaakt worden.
+        Inloggen kan door op de knop 'inloggen' te klikken zonder username of wachtwoord te gebruiken.
+
+        Dit is bij wijze van test even gedaan om 1 view te maken (die van de klusjesman)
+
+        TODO navigatie zo maken dat username en wachtwoord weer nodig zijn
+         */
         Column(
             modifier = Modifier
                 .fillMaxWidth()
